@@ -1,16 +1,15 @@
 package guru.springframework.spring5webapp;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
-public class Spring5webappApplicationTests {
-
+class Spring5webappApplicationTests
+{
 	@Test
-	public void contextLoads() {
+	void contextLoads()
+	{
 	}
 
 }
+
